refactor layouts y titulo listado general de reportes
