+++ resources/js/app.js
@@ -3,6 +3,7 @@ import 'flowbite';
 import initSelect from '@preline/select';
 import Alpine from 'alpinejs';
 import Chart from 'chart.js/auto';
+import './flowbite-datepicker';
 
 import.meta.glob(['../img/**']);
