+++ resources/views/layouts/guest.blade.php
@@ -1,19 +1,14 @@
 <!DOCTYPE html>
 <html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="csrf-token" content="{{ csrf_token() }}" />
-
-    <title>{{ config('app.name', 'Laravel') }}</title>
-
-    <!-- Fonts -->
+    <title>{{ config('app.name', 'ReportFIA') }}</title>
     <link rel="preconnect" href="https://fonts.bunny.net" />
     <link href="https://fonts.bunny.net/css?family=figtree:400,500,600&display=swap" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/notyf@3/notyf.min.css">
-    <!-- Scripts -->
     @vite(['resources/css/app.css', 'resources/js/app.js'])
 </head>
 
@@ -43,7 +38,6 @@
     </div>
     <script type="text/javascript" src="{{ asset('assets/js/jquery.min.js') }}"></script>
     <script src="https://cdn.jsdelivr.net/npm/notyf@3/notyf.min.js"></script>
-
     <script>
         var notyf = new Notyf({
             types: [{
@@ -88,7 +82,6 @@
                 },
             ]
         });
-
         const noty = (content, type = 'info') => {
             notyf.open({
                 type: type,
@@ -97,7 +90,6 @@
                 dismissible: true
             });
         }
-
         @if (!empty(session()->has('message')))
             noty(@json(session('message')['content']), @json(session('message')['type']) ?? 'success');
         @endif
@@ -114,25 +106,17 @@
     <script>
         document.addEventListener('DOMContentLoaded', function(e) {
             const loader = document.getElementById('loader');
-
             window.addEventListener('beforeunload', function() {
-                // Mostrar el loader al salir de la página
                 loader.classList.remove('hidden');
             });
-
             document.addEventListener('submit', function() {
-                // Mostrar el loader al enviar un formulario
                 loader.classList.remove('hidden');
-
                 if (event.defaultPrevented) {
-                    // Si el evento fue prevenido, ocultar el loader
                     loader.classList.add('hidden');
                 }
             });
-
-            // Ocultar el loader al regresar a la página
             window.addEventListener('pageshow', function(event) {
-                if (event.persisted) { // Si la página está cargada desde caché
+                if (event.persisted) {
                     loader.classList.add('hidden');
                 }
             });
